ユーザー取得とチェックの共通化
ユーザー取得とチェックをconcerns.user_resource_controller.rbにまとめて、必要に応じてincludeする形にしました。
