+++ household_account_book_app/app/controllers/households_controller.rb
@@ -1,6 +1,5 @@
 class HouseholdsController < ApplicationController
-  before_action :authorize_user
-
+  include UserResourceController
   def index
     @household = Household.new
   end
@@ -18,18 +17,7 @@ class HouseholdsController < ApplicationController
 
   private
 
-  def user
-    @user ||= User.find(params[:user_id])
-  end
-
   def household_params
     params.require(:household).permit(:name, :date, :amount, :category_id, :transaction_type)
-  end
-
-  def authorize_user
-    return if user == current_user
-
-    flash[:notice] = 'アクセス権限がありません'
-    redirect_to root_path
   end
 end
